backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.16,<4.0
django-cors-headers>=4.9,<5.0
django-environ>=0.13,<1.0
psycopg2-binary>=2.9,<3.0
